Pretend it's UNIX everywhere
line endings were a problem to us. Those who are working on Windows
(not me, thanks god) sometimes have strange problem. Most of the code
makes the assumtption that the line are ended by a \n (LF) character,
but on windows line are ended using TWO character \r\n (CRLF).

Here we implement a reader (that wraps another reader) that skips every
\r. The implementation may not be the most efficient (reading byte by
byte) but it works, and it seems fast enough (the os.File reader maybe
have some sort of buffer... who knows).